refactor: refactor auth middleware and all that related
