+++ internal/auth/infra/service.go
@@ -1,4 +1,4 @@
-package auth
+package infra
 
 import (
 	"encoding/json"
